add class for formatting methods in the DTOs, adjust comments in JobResource and JobService fix issue in client when retrieving all available jobs
